Reference count fun!
Turns out the MenuItem doesn't retain the target, so we need to make sure that the "glue" classes (that are standing in for the Objective-C blocks of the original) stick around.

Solution:
Make them children of the coverLayer -- then they'll stick around until it goes away.

Problem with that Solution:
The "glue" classes can't retain the layer or dialog anymore because then we would have circlar references.  Fun!

Solution to Solution's problem: (eek!)
The layer should be the parent of the "glue" classes since that was part of the original solution.  (Paranoia makes me verify this via a tag and some further paranoid dynamic_cast asserts that probably really don't have to be there.)  And the dialog can be found by the tag the original Objective-C version was using.

But we do have to be really careful when we remove the layer from it's parent, since that line is likely to destroy the layer and all it's children (us) -- so make sure it doesn't go away until the we're done with it.

This is tested and seems to work.